+++ core/src/main/java/arez/ArezConfig.native.js
@@ -1,0 +1,3 @@
+// Ensure that @defines are included so that they can specified at compile time for the closure compiler
+/** @suppress {extraRequire} */
+let arez = goog.require('arez');